FEAT: Update service, controller and route for Update Request. Rename constants file
