+++ src/controllers/request.js
@@ -32,10 +32,11 @@ const updateRequest = async (req, res) => {
   const result = await RequestService.updateRequest(id, req.body);
 
   // TODO: Check this errors
-  if (result.err) {
-    const { results, statusCode } = formatHTTPResponse(400, result);
-    return res.status(statusCode).json({ results });
-  }
+  // if (result.err) {
+
+  //   const { results, statusCode } = formatHTTPResponse(400, result);
+  //   return res.status(statusCode).json({ results });
+  // }
 
   return res.status(200).send();
 };
